// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/PayRollComponentModel.ts
export class PayRollComponentModel{
    payRollComponentId: string;
    componentName: string
    isDeduction: boolean;
    isVariablePay: boolean;
    isArchived: boolean;
    timeStamp: any;
    employeeContributionPercentage: number;
    employerContributionPercentage: number;
    isVisible: boolean;
    relatedToContributionPercentage : boolean;
    isBands: boolean;
}